Add input functions to gadgets in crt (#10)
* Add input_u64 to BigUintPoly

* Add input_u64 to MontgomeryPoly

* Adding input_biguint to CRTPoly

* Fix tests in crt/mod.rs

* Change the LUT definition

* FIx isolate gadget

* Redesign input functions in gadget and add PackedCrtContext

* Add tests for PackedCrtPoly

---------

+++ src/circuit/gate.rs
@@ -33,7 +33,7 @@ pub enum PolyGateType {
     Sub,
     Mul,
     LargeScalarMul { scalar: Vec<BigUint> },
-    Rotate { shift: usize },
+    Rotate { shift: i32 },
     PubLut { lookup_id: usize },
 }
